requirements.txt
alembic==1.15.2
blinker==1.9.0
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
ExifRead==3.0.0
filelock==3.18.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.58.0
fpdf==1.7.2
fsspec==2025.3.2
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==25.0
Pillow==10.2.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
SQLAlchemy==2.0.41
sympy==1.14.0
torch==2.2.1
torchvision==0.17.1+cpu
typing_extensions==4.13.2
Werkzeug==3.1.3 